Improve file reader crate
