--- file_system/src/fs/mod.rs
@@ -1,3 +1,0 @@
-pub use file_reader::FileReader;
-
-mod file_reader;
